+++ releasenotes.md
@@ -5,7 +5,7 @@
 * A bug has been fixed whereby in some rare instances objects would get suspended in midair when the sector they were on lowered.
 * The last WAD to be opened now prepopulates the `File name` field in *DOOM Retro’s* WAD launcher.
 * The `r_shakescreen` CVAR has been renamed to `r_shake_damage` and is now `50%` by default.
-* The screen will now shake when the player is near enough to an exploding barrel or rocket. This feature may be disabled using the new `r_shake_explode` CVAR.
+* The screen will now shake when the player is near enough to an exploding barrel. This feature may be disabled using the new `r_shake_barrels` CVAR.
 * When both the `vid_showfps` and `vid_vsync` CVARs are `on`, the frames per second displayed will be red if less than the display’s refresh rate.
 * An `exec` CCMD has been implemented that allows a series of commands stored in a file to be executed as if they had been typed in the console directly.
 * Textures that have patches with negative offsets (such as `TEKWALL1` and `STEP2`) now appear correctly.